Move golang packages under /pkg as is convention
