+++ build-backend.sh
@@ -8,6 +8,6 @@ set -e
 PROJECT_DIR=$(basename ${PWD})
 
 GIT_TAG=${SOURCE_GIT_TAG:-$(git describe --always --tags HEAD)}
-LD_FLAGS="-w -X github.com/openshift/console/version.Version=${GIT_TAG}"
+LD_FLAGS="-w -X github.com/openshift/console/pkg/version.Version=${GIT_TAG}"
 
 CGO_ENABLED=0 go build -ldflags "${LD_FLAGS}" -o bin/bridge github.com/openshift/console/cmd/bridge
