+++ pkg/auth/session.go
