+++ pkg/auth/session_test.go
